+++ lib/phx_live_storybook/entries_validator.ex
@@ -13,6 +13,7 @@ defmodule PhxLiveStorybook.EntriesValidator do
     validate_entry_icon(path, entry)
     validate_entry_component(path, entry)
     validate_entry_function(path, entry)
+    validate_entry_container(path, entry)
     validate_attribute_list_type(path, entry)
     validate_attribute_ids(path, attributes)
     validate_attribute_types(path, attributes)
@@ -55,6 +56,12 @@ defmodule PhxLiveStorybook.EntriesValidator do
 
   defp validate_entry_function(file_path, entry) do
     validate_type!(file_path, entry.function, :function, "entry function must be a function")
+  end
+
+  defp validate_entry_container(file_path, entry) do
+    unless entry.container in ~w(nil div iframe)a do
+      compile_error!(file_path, "entry container must be either :div or :iframe")
+    end
   end
 
   defp validate_attribute_list_type(file_path, entry) do